added-photo
Added photo with the white leds added and the blue and red switched with yellow leds, and used different wheels (from wooded wheels to construx wheels, but really doesn't matter what wheels you use). And has a buzzer located on the back and a battery for portability and not connected via a USB cable.